pyproject.toml
[project]
name = "telegram-bot-project"
version = "0.1.0"
description = "Telegram Bot for Digital Products"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==13.15",
    "requests>=2.32.4",
]
